optimize unconstrained minimization problems with vectorization
- Replace for-loops with vectorized JAX operations in CHNRSNBM, ENGVAL2, HELIX, and JENSMP
- Use arctan2 instead of conditional logic in HELIX for better performance
- Precompute common terms in ENGVAL2 to avoid redundant calculations
- Remove unnecessary type conversions in CHNRSNBM and JENSMP
- Update optimization todo with successful optimizations

+++ sif2jax/cutest/_unconstrained_minimisation/chnrsnbm.py
@@ -42,7 +42,7 @@ class CHNRSNBM(AbstractUnconstrainedMinimisation):
         # Alpha values are determined by sin(i) + 1.5
         # i_indices is [1, 2, ..., n-1], representing i=2..n in 1-based AMPL notation
         # So we need sin(2), sin(3), ..., sin(n), which is sin(i_indices + 1)
-        alpha_vals = jnp.sin((i_indices + 1).astype(float)) + 1.5
+        alpha_vals = jnp.sin(i_indices + 1) + 1.5
 
         # Get x[i-1] and x[i] values
         x_i_minus_1 = y[i_indices - 1]  # y[0] to y[n-2]
